Fix support for Gradle's configuration cache
This commit fixes support for Gradle's configuration cache. Before this change,
the plugin would break if `--configuraton-cache` was used. The changes are not
trivial, but I think it's for the better.

The first category of changes is on the build itself. It was quite difficult
to work with, because it involved publishing to maven local, and it broke
up-to-date checking. The build was using bad practices like reaching out to
other project's state or using explicit `dependsOn`. The problem with these
is that they were mostly incorrect and redundant.

Therefore, this commit also reworks the build itself in order to be self
contained, without having to rely on `mavenLocal()`. For example:

- the examples jar is now properly built as part of testing the Gradle
plugin. It is transparently added to the `core` test classpath, using an
exported configuration
- the `test-suite` tests are now executed as part of the build, using
Gradle's TestKit
- the JDK8 tests are executed as part of the Gradle plugin tests

There is no need to publish artifacts to Maven Local, but there is the
ability to publish all artifacts to a local Maven repository by running:

`./gradlew publishAllPublicationsToBuildRepository`

in which case the artifacts are published to the root `build/myRepo`
directory. This makes it much easier to check if the artifacts are
correctly published. This will include the Gradle plugin artifacts.

In addition, the manifest changes are only included for release,
or when a special flag is passed, in order to preserve up-to-date
checks and build caching.

I have tested this on a test Micronaut project with success.